Update: extract a few functions as library
